:open_file_folder: refactor: Move `utils` to `pkg` public package
