+++ remy/cycle_detector_injector_test.go
@@ -2,7 +2,7 @@ package remy
 
 import (
 	"fmt"
-	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
+	"github.com/wrapped-owls/goremy-di/remy/pkg/utils"
 	"testing"
 )
