+++ remy/internal/injector/standard_injector.go
@@ -2,7 +2,7 @@ package injector
 
 import (
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
-	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
+	"github.com/wrapped-owls/goremy-di/remy/pkg/utils"
 )
 
 type (
